+++ source/expression.cpp
@@ -2,8 +2,8 @@
 
 using namespace std;
 
-int calculateExpression(const string& expression) {
-    stack<int> numbers;
+float calculateExpression(const string& expression) {
+    stack<float> numbers;
     stack<char> operators;
     stack<char> parentheses;  // Stack per gestire le parentesi
 
@@ -13,7 +13,8 @@ int calculateExpression(const string& expression) {
         return 0;
     };
 
-    auto applyOperator = [](int num1, int num2, char op) {
+   // Dentro la funzione applyOperator, cambia il risultato da int a float
+    auto applyOperator = [](float num1, float num2, char op) {
         switch (op) {
             case '+':
                 return num1 + num2;
@@ -22,7 +23,7 @@ int calculateExpression(const string& expression) {
             case '*':
                 return num1 * num2;
             case '/':
-                if (num2 != 0) {
+                if (num2 != 0.0) { // Cambia da 0 a 0.0 per la divisione in virgola mobile
                     return num1 / num2;
                 } else {
                     cerr << ANSI_COLOR_RED << "Errore: Divisione per zero" << ANSI_COLOR_RESET << endl;
@@ -32,11 +33,11 @@ int calculateExpression(const string& expression) {
                 cerr << ANSI_COLOR_RED << "Errore: Operatore non valido" << ANSI_COLOR_RESET << endl;
                 exit(1);
         }
-    };
+    }; 
 
     for (size_t i = 0; i < expression.length(); i++) {
         if (isdigit(expression[i])) {
-            int num = 0;
+            float num = 0.0;
             while (i < expression.length() && isdigit(expression[i])) {
                 num = num * 10 + (expression[i] - '0');
                 i++;
@@ -46,15 +47,15 @@ int calculateExpression(const string& expression) {
             numbers.push(num);
         } else if (expression[i] == '+' || expression[i] == '-' || expression[i] == '*' || expression[i] == '/') {
             while (!operators.empty() && precedence(operators.top()) >= precedence(expression[i])) {
-                int num2 = numbers.top();
+                float num2 = numbers.top();
                 numbers.pop();
-                int num1 = numbers.top();
+                float num1 = numbers.top();
                 numbers.pop();
 
                 char op = operators.top();
                 operators.pop();
 
-                int result = applyOperator(num1, num2, op);
+                float result = applyOperator(num1, num2, op);
                 numbers.push(result);
             }
             operators.push(expression[i]);
@@ -65,12 +66,12 @@ int calculateExpression(const string& expression) {
                 char op = operators.top();
                 operators.pop();
 
-                int num2 = numbers.top();
+                float num2 = numbers.top();
                 numbers.pop();
-                int num1 = numbers.top();
+                float num1 = numbers.top();
                 numbers.pop();
 
-                int result = applyOperator(num1, num2, op);
+                float result = applyOperator(num1, num2, op);
                 numbers.push(result);
             }
 
@@ -84,15 +85,15 @@ int calculateExpression(const string& expression) {
     }
 
     while (!operators.empty()) {
-        int num2 = numbers.top();
+        float num2 = numbers.top();
         numbers.pop();
-        int num1 = numbers.top();
+        float num1 = numbers.top();
         numbers.pop();
 
         char op = operators.top();
         operators.pop();
 
-        int result = applyOperator(num1, num2, op);
+        float result = applyOperator(num1, num2, op);
         numbers.push(result);
     }
